The terminfo database uses i16 but ncurses uses c_int for variables.
We're just going to use i32 as this doesn't appear to be documented
anywhere and c_int is 32bits on most platforms.

+++ src/terminfo/mod.rs
@@ -153,7 +153,7 @@ impl<T: Write+Send> Terminal for TerminfoTerminal<T> {
     fn fg(&mut self, color: color::Color) -> io::Result<bool> {
         let color = self.dim_if_necessary(color);
         if self.num_colors > color {
-            return self.apply_cap("setaf", &[Param::Number(color as i16)]);
+            return self.apply_cap("setaf", &[Param::Number(color as i32)]);
         }
         Ok(false)
     }
@@ -161,7 +161,7 @@ impl<T: Write+Send> Terminal for TerminfoTerminal<T> {
     fn bg(&mut self, color: color::Color) -> io::Result<bool> {
         let color = self.dim_if_necessary(color);
         if self.num_colors > color {
-            return self.apply_cap("setab", &[Param::Number(color as i16)]);
+            return self.apply_cap("setab", &[Param::Number(color as i32)]);
         }
         Ok(false)
     }
